# package/backend/dist/swagger.json
{
	"components": {
...
		"schemas": {
			"UserCreatedDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"nickname",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserDTO": {
				"properties": {
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"nickname",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginResponseDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginUserDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseDTO": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgetPassDTO": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePassDTO": {
				"properties": {
					"password": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nickname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"isValid": {
						"type": "boolean"
					},
					"roleId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"$ref": "#/components/schemas/RoleDTO"
					}
				},
				"required": [
					"id",
					"nickname",
					"email",
					"image",
					"isValid",
					"roleId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"imageURL": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"imageURL"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IValidateErrorJSON": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateRoleRequestDTO": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ThreadDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"category": {
						"$ref": "#/components/schemas/CategoryDTO"
					},
					"user": {
						"$ref": "#/components/schemas/UserDTO"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"content",
					"category",
					"user",
					"tags",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreadPaginatedDTO": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ThreadDTO"
						},
						"type": "array"
					},
					"lastPage": {
						"type": "number",
						"format": "double"
					},
					"totalRecords": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"hasMorePages": {
						"type": "boolean"
					}
				},
				"required": [
					"data",
					"lastPage",
					"totalRecords",
					"currentPage",
					"hasMorePages"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "NodeJS API for LosDelFondo WebApp",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCreatedDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUserDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/forgetPassword": {
			"post": {
				"operationId": "ForgetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgetPassDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/changePassword/{id}": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePassDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/v1/roles/{id}": {
			"get": {
				"operationId": "GetRoleByID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateErrorJSON"
								}
							}
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleRequestDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteRole",
				"responses": {
					"200": {
						"description": "Role Deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateErrorJSON"
								}
							}
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/roles": {
			"get": {
				"operationId": "GetAllRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RoleDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateErrorJSON"
								}
							}
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RoleDTO"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateErrorJSON"
								}
							}
						}
					}
				},
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleRequestDTO"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/threads": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadPaginatedDTO"
								}
							}
						}
					},
					"404": {
						"description": "No threads found"
					}
				},
				"tags": [
					"Thread"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "author",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadDTO"
								}
							}
						}
					},
					"201": {
						"description": "Success"
					},
					"400": {
						"description": "Validation Error"
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"tags": [
					"Thread"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/api/v1/threads/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ThreadDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Thread"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					},
					"403": {
						"description": "Thread not owner error"
					},
					"404": {
						"description": "Thread not found error"
					}
				},
				"tags": [
					"Thread"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseDTO"
								}
							}
						}
					},
					"403": {
						"description": "Thread not owner error"
					}
				},
				"tags": [
					"Thread"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}